WIP PoC: use state library for form state
To avoid that all form fields's React components are called whenever any
form state currently placed in React context changes. The library
`use-context-selector` is preventing re-renders, but not function calls,
as discussed for example in
https://github.com/dai-shi/use-context-selector/issues/92. Unfortunately
the function calls are CPU-heavy enough that on systems with CPUs that
are not top of the line, the form's UI responsiveness is severely
hampered by the form field component calls on any form context change.
The author of `use-context-selector` recommends the popular React state
management library `zustand` for cases where more fine-grained
reactivity is required.

This commit is a mini WIP PoC of using Zustand instead for React context
in the form components. Currently it's only implementing the `UPDATE`
dispatch to showcase that form inputs work with little effort.

The main cost of completing this migration would be to migrate other
form related contexts which currently force any child of
`Form/index.tsx` to re-render on any context change.